requirements.txt
flask==3.0.0
pandas==2.1.4
numpy==1.26.2
plotly==5.18.0
werkzeug==3.0.1
google-generativeai==0.3.2
python-dotenv==1.0.1
flask-cors==4.0.0
